+++ src/main/java/org/toughradius/entity/Bras.java
@@ -1,12 +1,14 @@
 package org.toughradius.entity;
 
+import sun.rmi.runtime.Log;
+
 import java.io.Serializable;
 import java.sql.Timestamp;
 import java.util.Date;
 
 public class Bras implements Serializable{
 
-    private Integer id;
+    private Long id;
 
     private String identifier;
 
@@ -30,11 +32,11 @@ public class Bras implements Serializable{
 
     private Timestamp createTime;
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
